Update side_ch scripts for this discussion: https://github.com/open-sdr/openwifi/discussions/344

+++ user_space/side_ch_ctl_src/iq_capture.py
@@ -55,7 +55,7 @@ def parse_iq(iq, iq_len):
     iq = iq.reshape([num_trans, num_int16_per_trans])
     
     timestamp = iq[:,0] + pow(2,16)*iq[:,1] + pow(2,32)*iq[:,2] + pow(2,48)*iq[:,3]
-    iq_capture = iq[:,4::4] + iq[:,5::4]*1j
+    iq_capture = np.int16(iq[:,4::4]) + np.int16(iq[:,5::4])*1j
     agc_gain = iq[:,6::4]
     rssi_half_db = iq[:,7::4]
     # print(num_trans, iq_len, iq_capture.shape, agc_gain.shape, rssi_half_db.shape)
@@ -106,7 +106,7 @@ while True:
         if (test_residual != 0):
             print("Abnormal length")
 
-        iq = np.frombuffer(data, dtype='int16')
+        iq = np.frombuffer(data, dtype='uint16')
         np.savetxt(iq_fd, iq)
 
         timestamp, iq_capture, agc_gain, rssi_half_db = parse_iq(iq, iq_len)
